+++ src/logic/getCheckboxValue.ts
@@ -1,7 +1,7 @@
 import isArray from '../utils/isArray';
 import isUndefined from '../utils/isUndefined';
 import isEmptyString from '../utils/isEmptyString';
-import { RadioOrCheckboxOption } from '../types/types';
+import { RadioOrCheckboxOption } from '../types/form';
 
 type CheckboxFieldResult = {
   isValid: boolean;
