+++ src/logic/getFieldArrayValueByName.ts
@@ -1,7 +1,7 @@
 import transformToNestObject from './transformToNestObject';
 import getFieldsValues from './getFieldsValues';
 import get from '../utils/get';
-import { FieldValues, FieldRefs, InternalFieldName } from '../types/types';
+import { FieldValues, FieldRefs, InternalFieldName } from '../types/form';
 
 export default <TFieldValues extends FieldValues = FieldValues>(
   fields: FieldRefs<TFieldValues>,
