+++ src/useFormContext.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { UseFormMethods, FieldValues } from './types/types';
+import { UseFormMethods, FieldValues } from './types/form';
 import { FormProviderProps } from './types/props';
 
 export const FormContext = React.createContext<UseFormMethods | null>(null);
