+++ src/utils/isRadioInput.ts
@@ -1,4 +1,4 @@
-import { FieldElement } from '../types/types';
+import { FieldElement } from '../types/form';
 
 export default (element: FieldElement): element is HTMLInputElement =>
   element.type === 'radio';
