+++ src/utils/isSelectInput.ts
@@ -1,4 +1,4 @@
-import { FieldElement } from '../types/types';
+import { FieldElement } from '../types/form';
 import { SELECT } from '../constants';
 
 export default (element: FieldElement): element is HTMLSelectElement =>
